Allowed multiple enumerations.
